New: Basic pykka messaging.
Added methods to handle and keep track of running GameInstances. Have implemented some basic query parameters that can be passed in like so /game/[id]/?[query_parameter]=[value] to start, stop and interact with pykka actors.

Right now, you can take a game_instance, start it, tell it to do some turns (all this does for the moment is increment the turn counter) and stop it.

It's not much, but its a big step in the right direction (and a big proof of concept for how we planned to wire this all up).