Fixed: Brought WorldState in line with expected API.
Armin, I'm not sure if you'll think I'm a hero or a monster, but whats done is done.

Essentially I hacked together all the functionality of the old world implementation... Calling world[x][y] will return a list which includes cells and actors and other inhabitants at that location. This is one of those things where I'm not sure if I've done something beautiful or terrible... There's a private class involved.

In any case, it's all very easy to use now!