New: Added concept of 'master' actor models.
Creating an actor in the DB will create a 'master', which is the definitive representation of the actors name and behaviour. Whenever you modify actors, you are modifying the masters. Adding an actor to a game will create a copy of the actor. This will make it easy to create an actor, then add it to multiple games.