New: Big game_runner feature additions.
Added pausing/resuming: Game will refuse edits if not paused. Game starts paused.
Added 'spawn' deltas: When an actor is added to a game, it will create a spawn delta signifiying when the actor was added.
Added rollbacks to commands: 'stop' and 'pause' require an 'on_turn' argument, that will be used to roll back any pre-computed changes on the backend. This is so that stopping the game or pausing and editing at a particular turn will apply the changes at the exact turn the user paused on.